+++ src/threads.js
@@ -2008,6 +2008,35 @@ Process.prototype.reportItems = function (indices, list) {
 
 // Process - other basic list accessors
 
+Process.prototype.reportTableColumn = function (index, list) {
+    // experimental and probably controversial as a primitive,
+    // because it's so nice and easy to write in Snap!
+    var col;
+
+    function columns() {
+        return Math.max(...list.itemsArray().map(row =>
+            row instanceof List ? row.length() : 0)
+        );
+    }
+
+    if (!this.isMatrix(list)) {
+        throw new Error(
+            'expecting ' + 'table' + ' but getting ' + this.reportTypeOf(list)
+        );
+    }
+    if (index === '') {
+        return new List(new Array(list.length()));
+    }
+    if (this.inputOption(index) === 'any') {
+        col = this.reportBasicRandom(1, columns());
+        return list.map(row => row.at(col));
+    }
+    if (this.inputOption(index) === 'last') {
+        return list.map(row => row.at(columns()));
+    }
+    return list.map(row => row.at(index));
+};
+
 Process.prototype.reportListLength = function (list) {
     this.assertType(list, 'list');
     return list.length();